requirements.txt
streamlit
azure-devops